fix: harden iOS note swipe actions
