+++ apps/desktop/src/mobile/note-actions-menu.test.tsx
@@ -3,7 +3,7 @@ import { render } from 'vitest-browser-react'
 import { page } from 'vitest/browser'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import { setBridge } from '@reflect/core'
+import { setBridge, type GraphInfo } from '@reflect/core'
 import { NoteActionsMenu } from './note-actions-menu'
 
 /**
@@ -24,9 +24,10 @@ vi.mock('@/components/ui/drawer', () => ({
   DrawerTitle: ({ children }: { children?: ReactNode }) => <h2>{children}</h2>,
 }))
 
-vi.mock('@/providers/graph-provider', () => ({
-  useGraph: () => ({ graph: { root: '/g', name: 'g', generation: 1 } }),
+const graphState = vi.hoisted<{ current: GraphInfo | null }>(() => ({
+  current: { root: '/g', name: 'g', generation: 1 },
 }))
+vi.mock('@/providers/graph-provider', () => ({ useGraph: () => ({ graph: graphState.current }) }))
 // The pinned set comes from the index; an empty list means "not pinned".
 vi.mock('@/hooks/use-pinned-notes', () => ({ usePinnedNotes: () => [] }))
 // No session is open for the note in this unit; discard is a no-op lookup.
@@ -38,6 +39,7 @@ const mockInvoke = vi.fn<(command: string, args: Record<string, unknown>) => Pro
 setBridge({ invoke: mockInvoke, listen: async () => () => {} })
 
 beforeEach(() => {
+  graphState.current = { root: '/g', name: 'g', generation: 1 }
   calls.length = 0
   mockInvoke.mockReset()
   mockInvoke.mockImplementation(async (command, args) => {
@@ -86,5 +88,18 @@ describe('NoteActionsMenu', () => {
       expect(calls.some((call) => call.command === 'note_delete')).toBe(true)
     })
     expect(onDeleted).toHaveBeenCalledOnce()
+  })
+
+  it('reports when the graph disappears before delete confirmation', async () => {
+    graphState.current = null
+    const { view, onDeleted } = await mount()
+
+    await view.getByRole('button', { name: 'Delete' }).click()
+    const dialog = page.getByRole('dialog')
+    await dialog.getByRole('button', { name: 'Delete' }).click()
+
+    await expect.element(dialog.getByText('No graph is open.')).toBeInTheDocument()
+    expect(calls.some((call) => call.command === 'note_delete')).toBe(false)
+    expect(onDeleted).not.toHaveBeenCalled()
   })
 })
